pyproject.toml
[project]
name = "quicktask"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "psycopg[binary]>=3.3.3",
    "pytelegrambotapi>=4.33.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
]
